29.08.2021

+++ src/main/java/amino/models/requests/CheckInBody.java
@@ -4,6 +4,5 @@ import amino.utils.TimeUtils;
 
 public class CheckInBody {
     private String timezone = "120";
-    private Integer timestamp = TimeUtils.unixInstant();
+    private Long timestamp = TimeUtils.unixInstant();
 }
-
