+++ src/main/java/amino/models/requests/PlayLotteryBody.java
@@ -4,6 +4,6 @@ import amino.utils.TimeUtils;
 
 public class PlayLotteryBody {
     private String timezone = "120";
-    private Integer timestamp = TimeUtils.unixInstant();
+    private Long timestamp = TimeUtils.unixInstant();
 }
