+++ src/main/java/amino/models/requests/SendCoinsBody.java
@@ -1,12 +1,11 @@
 package amino.models.requests;
 
-import amino.models.requests.TippingContext;
 import amino.utils.TimeUtils;
 
 public class SendCoinsBody {
     private Integer coins;
     private TippingContext tippingContext;
-    private Integer timestamp = TimeUtils.unixInstant();
+    private Long timestamp = TimeUtils.unixInstant();
 
     public SendCoinsBody coins(Integer coins) {
         this.coins = coins;
